# Conflicts:
#	.gitignore
#	FeedBuilder.config
#	HotsBpHelper.sln
#	HotsBpHelper/.config/AppSetting.json
#	HotsBpHelper/Api/RestApi.cs
#	HotsBpHelper/App.config
#	HotsBpHelper/App.xaml
#	HotsBpHelper/App.xaml.cs
#	HotsBpHelper/Bootstrapper.cs
#	HotsBpHelper/Const.cs
#	HotsBpHelper/HotsBpHelper.csproj
#	HotsBpHelper/Localization/Strings.Designer.cs
#	HotsBpHelper/Localization/Strings.en-US.resx
#	HotsBpHelper/Localization/Strings.zh-CN.resx
#	HotsBpHelper/Localization/Strings.zh-TW.resx
#	HotsBpHelper/Pages/BpView.xaml
#	HotsBpHelper/Pages/BpView.xaml.cs
#	HotsBpHelper/Pages/BpViewModel.cs
#	HotsBpHelper/Pages/HeroSelectorViewModel.cs
#	HotsBpHelper/Pages/MapSelectorView.xaml
#	HotsBpHelper/Pages/MapSelectorViewModel.cs
#	HotsBpHelper/Pages/SelectorViewModel.cs
#	HotsBpHelper/Pages/ShellView.xaml
#	HotsBpHelper/Pages/ShellViewModel.cs
#	HotsBpHelper/Pages/ViewModelBase.cs
#	HotsBpHelper/Properties/AssemblyInfo.cs
#	HotsBpHelper/Settings/AppSetting.cs
#	HotsBpHelper/UserControls/WebKitBrowser.xaml.cs
#	HotsBpHelper/Utils/ComboBoxItemUtil/HeroItemUtil.cs
#	HotsBpHelper/Utils/ComboBoxItemUtil/MapItemUtil.cs
#	HotsBpHelper/Utils/ImageUtils.cs
#	HotsBpHelper/Utils/ScreenUtil.cs
#	HotsBpHelper/WPF/FilteredComboBox.cs
#	HotsBpHelper/packages.config

+++ HotsBpHelper/Utils/ScreenUtil.cs
@@ -210,7 +210,8 @@ namespace HotsBpHelper.Utils
         public static Point ToPixelPoint(this Point unitPoint)
         {
             int pixelX, pixelY;
-            TransformToPixels((int)unitPoint.X, (int)unitPoint.Y, out pixelX, out pixelY);
+
+            TransformToPixels(unitPoint.X, unitPoint.Y, out pixelX, out pixelY);
             return new Point(pixelX, pixelY);
         }
 
@@ -222,7 +223,7 @@ namespace HotsBpHelper.Utils
         /// <param name="pixelY">returns the Y value in pixels</param>
         /// <param name="unitX">a device independent unit value X</param>
         /// <param name="unitY">a device independent unit value Y</param>
-        public static void TransformFromPixels(int pixelX, int pixelY, out double unitX, out double unitY)
+        public static void TransformFromPixels(int pixelX, int pixelY, out int unitX, out int unitY)
         {
             IntPtr hDc = GetDC(IntPtr.Zero);
             if (hDc != IntPtr.Zero)
@@ -231,9 +232,9 @@ namespace HotsBpHelper.Utils
                 int dpiY = GetDeviceCaps(hDc, (int)DeviceCap.LOGPIXELSY);
 
                 ReleaseDC(IntPtr.Zero, hDc);
-
-                unitX = pixelX * 96 / (double)dpiX;
-                unitY = pixelY * 96 / (double)dpiY;
+                
+                unitX = (int) (pixelX * 96 / (double)dpiX);
+                unitY = (int) (pixelY * 96 / (double)dpiY);
             }
             else
                 throw new ArgumentNullException("Failed to get DC.");
@@ -241,16 +242,16 @@ namespace HotsBpHelper.Utils
 
         public static Point ToUnitPoint(this Point pixelPoint)
         {
-            double unitX, unitY;
-            TransformFromPixels((int)pixelPoint.X, (int)pixelPoint.Y, out unitX, out unitY);
-            return new Point((int)unitX, (int)unitY);
+            int unitX, unitY;
+            TransformFromPixels(pixelPoint.X, pixelPoint.Y, out unitX, out unitY);
+            return new Point(unitX, unitY);
 
         }
         public static Size ToUnitSize(this Size pixelSize)
         {
-            double unitWidth, unitHeight;
-            TransformFromPixels((int)pixelSize.Width, (int)pixelSize.Height, out unitWidth, out unitHeight);
-            return new Size((int)unitWidth, (int)unitHeight);
+            int unitWidth, unitHeight;
+            TransformFromPixels(pixelSize.Width, pixelSize.Height, out unitWidth, out unitHeight);
+            return new Size(unitWidth, unitHeight);
 
         }
